requirements.txt
matplotlib==3.6.2
numpy==1.24.0
pandas==1.5.2
pysftp==0.2.9
python-dotenv==0.21.0